Searching: Split out search tests into their own dir
